+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 - General:
   - Native file dialogs (awt)
   - Project panel now opens with a click as default, you can roll back to hovering in settings.
-  - Project pane can be dragged around
+  - Project pane can be dragged around and better description
 - Grids:
   - Grid view can be forced if header detection result is uncertain
   - Grid columns automatic initial widths
   - Grid view font is the same as in text viewer
 - External scripts:
   - Support for search output sections (specify a `sectionRegex` in *.jloga.json* file)
+- Bug fixes:
+  - Unselectable values in project combos
 
 ## v0.2.3
 - General: